navbar_place_categories_search: move place/category store to search
