--- frontend/src/store/modules/place/category/index.js
@@ -1,10 +1,0 @@
-import state from './state';
-import actions from './actions';
-import mutations from './mutations';
-
-export default {
-    namespaced: true,
-    state,
-    actions,
-    mutations
-};
