--- frontend/src/store/modules/place/category/mutations.js
@@ -1,5 +1,0 @@
-export default {
-    SET_SEARCH_CATEGORIES: (state, searchCategories) => {
-        state.searchCategories = searchCategories;
-    }
-};
