--- frontend/src/store/modules/place/category/state.js
@@ -1,3 +1,0 @@
-export default {
-    searchCategories: []
-};
