+++ frontend/src/store/modules/search/state.js
@@ -1,3 +1,4 @@
 export default {
-    city: {}
+    city: {},
+    placeCategory: {},
 };
